+++ LookupAnything/Framework/DebugFields/GenericDebugField.cs
@@ -1,4 +1,4 @@
-﻿using System.Globalization;
+using System.Globalization;
 
 namespace Pathoschild.Stardew.LookupAnything.Framework.DebugFields
 {
@@ -8,17 +8,20 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.DebugFields
         /*********
         ** Accessors
         *********/
-        /// <summary>A short field label.</summary>
+        /// <inheritdoc />
         public string Label { get; protected set; }
 
-        /// <summary>The field value.</summary>
+        /// <inheritdoc />
         public string Value { get; protected set; }
 
-        /// <summary>Whether the field should be displayed.</summary>
+        /// <inheritdoc />
         public bool HasValue { get; protected set; }
 
-        /// <summary>Whether the field should be highlighted for special attention.</summary>
+        /// <inheritdoc />
         public bool IsPinned { get; protected set; }
+
+        /// <inheritdoc />
+        public string OverrideCategory { get; set; }
 
 
         /*********