+++ LookupAnything/Framework/Lookups/Tiles/TileSubject.cs
@@ -77,12 +77,20 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Tiles
         /// <summary>Get raw debug data to display for this subject.</summary>
         public override IEnumerable<IDebugField> GetDebugFields()
         {
+            string mapTileLabel = I18n.Type_MapTile();
+            string locationLabel = I18n.Tile_GameLocation();
+
+            // tiles
             Tile[] tiles = this.GetTiles(this.Location, this.Position).ToArray();
             foreach (Tile tile in tiles)
             {
                 foreach (IDebugField field in this.GetDebugFieldsFrom(tile))
-                    yield return new GenericDebugField($"{tile.Layer.Id}::{field.Label}", field.Value, field.HasValue, field.IsPinned);
+                    yield return new GenericDebugField($"{tile.Layer.Id}::{field.Label}", field.Value, field.HasValue) { OverrideCategory = mapTileLabel };
             }
+
+            // location
+            foreach (IDebugField field in this.GetDebugFieldsFrom(this.Location))
+                yield return new GenericDebugField(field.Label, field.Value, field.HasValue, field.IsPinned) { OverrideCategory = locationLabel };
         }
 
         /// <summary>Draw the subject portrait (if available).</summary>